Trac #26623: add cones.<whatever> shortcuts for common cones.
Add a new module sage.geometry.cones (src/sage/geometry/cones.py) that
allows users to easily construct a few polyhedral cones:

  * The nonnegative orthant
  * The rearrangement cone
  * The Schur cone
  * The trivial cone

Inside the new module, a "ConvexRationalPolyhedralConeFactory" class
is defined. The class is not meant to be instantiated; instead, one
pre-existing instance, called "cones", is created for the user. Using
it is as simple as

  sage: cones.nonnegative_orthant(3)

This functionality is in the same spirit as the pre-defined "graphs"
object that lets one create the Petersen graph with

  sage: graphs.PetersenGraph()

After adding the new module, the "cones" object is imported into
sage.geometry.all to ensure that it is available to users immediately
as in the above example.