Trac #26623: factor out common "cones" argument processing.
There's a lot of duplication in the ConvexRationalPolyhedralConeFactory
class. Sadly, most of it is in the documentation, and I don't know
what to do about that. But, one thing we can fix: each cones.<whatever>
method currently performs exactly the same steps on its ambient_dim and
lattice arguments, to ensure that they aren't "None", and that they are
compatible with one another (errors are raised otherwise).

Since that pre-processing is exactly the same for every method, this
commit factors it out into a _preprocess_args() method. The new method
should not appear in the tab-completion list for the "cones" object.