Trac #28958: spkg-configure.m4 for nauty.
The Nauty package has no dependencies, and SageMath utilizes it only
through a command-line (popen) interface. The following three files
call Nauty executables,

  1. src/sage/graphs/digraph_generators.py (directg, gentourng)
  2. src/sage/graphs/graph_generators.py (geng)
  3. src/sage/graphs/hypergraph_generators.py (genbg)

and those are the only four executables used. They are called by name,
without a path, and are thus presumed to be in the process's PATH.

This commit adds an spkg-configure.m4 file to detect a usable copy of
Nauty on the system. Per the discussion above, it does so by checking
for those four executables in the user's PATH.