Trac #28958: allow tests to pass with nauty-2.7.x.
We have Trac #26891 open to upgrade our version of Nauty from v2.6 to v2.7.
The "geng" program gains some additional (optional) arguments in v2.7,

  http://pallini.di.uniroma1.it/changes24-27.txt

and as a result, one of our doctests needs to be updated. Currently,
it looks for the exact "usage" string output by geng. In preparation
for detecting a system copy of nauty, this commit updates the expected
usage string to match either the old output from nauty-2.6, or the new
output from nauty-2.7. This will make the upgrade to nauty-2.7 a bit
easier, and means that we don't have to worry about detecting a
specific version of nauty on the system; the remaining tests pass
with either nauty-2.6 or nauty-2.7 installed.