Trac #28958: make the nauty check more robust.
On Debian and Ubuntu at least, the nauty executables are all prefixed
with "nauty-", so you wind up with /usr/bin/nauty-geng, for example.
To ensure that our spkg-configure.m4 can detect this requires some work.

First, the paths to the executables were factored out into environment
variables. Those paths are imported in the files that popen() them,
and two tests were modified to ignore a prefixed executable (these are
bad tests, in my opinion, but it was easier to fix them than to stop
and rewrite them).

Then, the spkg-configure.m4 for nauty was modified to search for both
the prefixed and un-prefixed executables, storing the results in four
environment variables. If we wind up using the system copy of nauty,
those variables are left alone. Otherwise, we revert them to the bare
executable names like "directg" and "geng" that the nauty spkg expects.
Finally, we substitute the values of those variables into the file
src/bin/sage-env-config, from whence they are imported.

This is tested to work with a mix of prefixed and unprefixed variables,
and with --with-system-nauty=no.